Mining: support mutable "completeness" oracle.
Requested change: support for a mutable definition of "mission is complete." Default behavior marks "complete" missions as having at least one each of lander, landing site, and payload.

Motivating idea: allow "payloads" to represent parts of the lander (solar panels, tanks, engines, legs) rather than cargo. A mission is "complete" if the lander is fully assembled based on staked components.

Implementing typing data on the Payloads token is being left as a post-deployment feature. Support for that feature is added here, in the form of a settable "completeMissionChecker" that evaluates tokens and indicates if they represent a complete mission.

Note: the scope of code changes to IMSpaceMissionMining.sol here exceed the addition of this feature. This is because adding the feature pushed the compiled contract size beyond Ethereum network limitations, which may be reproduced on other chains. The contract needed to be further optimized to bring its compiled size back below the limit.